- Fix CONFIG_MODULES=n case. - config.conf: Some cleanups.

+++ rpm/post.sh
@@ -26,27 +26,29 @@ if [ -f /proc/cpuinfo ]; then
     esac
 fi
 
-echo Setting up /lib/modules/@KERNELRELEASE@
-suffix=
-case @FLAVOR@ in
-    kdump|ps3|um|xen*)
-	suffix=-@FLAVOR@
-	;;
-esac
-for x in /boot/@IMAGE@ /boot/initrd; do
-    rm -f $x$suffix
-    ln -s ${x##*/}-@KERNELRELEASE@ $x$suffix
-done
+if [ -e /lib/modules/@KERNELRELEASE@ ]; then
+    echo Setting up /lib/modules/@KERNELRELEASE@
+    suffix=
+    case @FLAVOR@ in
+	kdump|ps3|um|xen*)
+	    suffix=-@FLAVOR@
+	    ;;
+    esac
+    for x in /boot/@IMAGE@ /boot/initrd; do
+	rm -f $x$suffix
+	ln -s ${x##*/}-@KERNELRELEASE@ $x$suffix
+    done
 
-if [ -x /sbin/module_upgrade ]; then
-    /sbin/module_upgrade --rename mptscsih="mptspi mptfc mptsas"
-fi
+    if [ -x /sbin/module_upgrade ]; then
+	/sbin/module_upgrade --rename mptscsih="mptspi mptfc mptsas"
+    fi
 
-# Add symlinks of compatible modules to /lib/modules/$krel/weak-updates/.
-if [ -x /usr/lib/module-init-tools/weak-modules ]; then
-    /usr/lib/module-init-tools/weak-modules --add-kernel @KERNELRELEASE@
+    # Add symlinks of compatible modules to /lib/modules/$krel/weak-updates/.
+    if [ -x /usr/lib/module-init-tools/weak-modules ]; then
+	/usr/lib/module-init-tools/weak-modules --add-kernel @KERNELRELEASE@
+    fi
+    /sbin/depmod -a -F /boot/System.map-@KERNELRELEASE@ @KERNELRELEASE@
 fi
-/sbin/depmod -a -F /boot/System.map-@KERNELRELEASE@ @KERNELRELEASE@
 
 message_install_bl () {
 	echo "You may need to setup and install the boot loader using the"
